refactor: split expression resolver
