+++ crates/sail-plan/src/resolver/expression/function.rs
@@ -1,0 +1,156 @@
+use datafusion_common::DFSchemaRef;
+use datafusion_expr::expr;
+use datafusion_expr::expr::ScalarFunction;
+use datafusion_expr::registry::FunctionRegistry;
+use sail_common::spec;
+use sail_python_udf::udf::pyspark_unresolved_udf::PySparkUnresolvedUDF;
+
+use crate::error::{PlanError, PlanResult};
+use crate::function::common::{AggFunctionInput, FunctionContextInput, ScalarFunctionInput};
+use crate::function::{get_built_in_aggregate_function, get_built_in_function};
+use crate::resolver::expression::NamedExpr;
+use crate::resolver::function::PythonUdf;
+use crate::resolver::state::PlanResolverState;
+use crate::resolver::PlanResolver;
+use crate::utils::ItemTaker;
+
+impl PlanResolver<'_> {
+    pub(super) async fn resolve_expression_function(
+        &self,
+        function: spec::UnresolvedFunction,
+        schema: &DFSchemaRef,
+        state: &mut PlanResolverState,
+    ) -> PlanResult<NamedExpr> {
+        let mut scope = state.enter_config_scope();
+        let state = scope.state();
+        let spec::UnresolvedFunction {
+            function_name,
+            arguments,
+            named_arguments,
+            is_distinct,
+            is_user_defined_function: _,
+            is_internal: _,
+            ignore_nulls,
+            filter,
+            order_by,
+        } = function;
+
+        let Ok(function_name) = <Vec<String>>::from(function_name).one() else {
+            return Err(PlanError::unsupported("qualified function name"));
+        };
+        if !named_arguments.is_empty() {
+            return Err(PlanError::todo("named function arguments"));
+        }
+        let canonical_function_name = function_name.to_ascii_lowercase();
+        if let Ok(udf) = self.ctx.udf(&canonical_function_name) {
+            if udf.inner().as_any().is::<PySparkUnresolvedUDF>() {
+                state.config_mut().arrow_allow_large_var_types = true;
+            }
+        }
+
+        let (argument_display_names, arguments) = self
+            .resolve_expressions_and_names(arguments, schema, state)
+            .await?;
+
+        // FIXME: `is_user_defined_function` is always false,
+        //   so we need to check UDFs before built-in functions.
+        let func = if let Ok(udf) = self.ctx.udf(&canonical_function_name) {
+            if ignore_nulls.is_some() || filter.is_some() || order_by.is_some() {
+                return Err(PlanError::invalid("invalid scalar function clause"));
+            }
+            if let Some(f) = udf.inner().as_any().downcast_ref::<PySparkUnresolvedUDF>() {
+                let function = PythonUdf {
+                    python_version: f.python_version().to_string(),
+                    eval_type: f.eval_type(),
+                    command: f.command().to_vec(),
+                    output_type: f.output_type().clone(),
+                };
+                self.resolve_python_udf_expr(
+                    function,
+                    &function_name,
+                    arguments,
+                    &argument_display_names,
+                    schema,
+                    f.deterministic(),
+                    is_distinct,
+                    state,
+                )?
+            } else {
+                expr::Expr::ScalarFunction(ScalarFunction {
+                    func: udf,
+                    args: arguments,
+                })
+            }
+        } else if let Ok(func) = get_built_in_function(&canonical_function_name) {
+            if ignore_nulls.is_some() || filter.is_some() || order_by.is_some() {
+                return Err(PlanError::invalid("invalid scalar function clause"));
+            }
+            let input = ScalarFunctionInput {
+                arguments,
+                function_context: FunctionContextInput {
+                    argument_display_names: &argument_display_names,
+                    plan_config: &self.config,
+                    session_context: self.ctx,
+                    schema,
+                },
+            };
+            func(input)?
+        } else if let Ok(func) = get_built_in_aggregate_function(&canonical_function_name) {
+            let filter = match filter {
+                Some(x) => Some(Box::new(self.resolve_expression(*x, schema, state).await?)),
+                None => None,
+            };
+            let order_by = match order_by {
+                Some(x) => self.resolve_sort_orders(x, true, schema, state).await?,
+                None => vec![],
+            };
+            let input = AggFunctionInput {
+                arguments,
+                distinct: is_distinct,
+                ignore_nulls,
+                filter,
+                order_by,
+                function_context: FunctionContextInput {
+                    argument_display_names: &argument_display_names,
+                    plan_config: &self.config,
+                    session_context: self.ctx,
+                    schema,
+                },
+            };
+            func(input)?
+        } else {
+            return Err(PlanError::unsupported(format!(
+                "unknown function: {function_name}",
+            )));
+        };
+
+        let name = self.config.plan_formatter.function_to_string(
+            &function_name,
+            argument_display_names.iter().map(|x| x.as_str()).collect(),
+            is_distinct,
+        )?;
+        Ok(NamedExpr::new(vec![name], func))
+    }
+
+    pub(super) async fn resolve_expression_call_function(
+        &self,
+        function_name: spec::ObjectName,
+        arguments: Vec<spec::Expr>,
+        schema: &DFSchemaRef,
+        state: &mut PlanResolverState,
+    ) -> PlanResult<NamedExpr> {
+        let function = spec::UnresolvedFunction {
+            function_name,
+            arguments,
+            named_arguments: vec![],
+            is_distinct: false,
+            is_user_defined_function: false,
+            is_internal: None,
+            ignore_nulls: None,
+            filter: None,
+            order_by: None,
+        };
+        self.resolve_expression_function(function, schema, state)
+            .await
+    }
+}
