+++ crates/sail-plan/src/resolver/function.rs
@@ -1,14 +1,6 @@
-use std::sync::Arc;
-
 use datafusion::arrow::datatypes::DataType;
-use datafusion_common::{plan_err, DFSchemaRef, DataFusionError, Result};
-use datafusion_expr::expr::AggregateFunctionParams;
-use datafusion_expr::{expr, AggregateUDF, Expr, ExprSchemable, ScalarUDF};
+use datafusion_common::plan_err;
 use sail_common::spec;
-use sail_python_udf::cereal::pyspark_udf::PySparkUdfPayload;
-use sail_python_udf::get_udf_name;
-use sail_python_udf::udf::pyspark_udaf::PySparkGroupAggregateUDF;
-use sail_python_udf::udf::pyspark_udf::{PySparkUDF, PySparkUdfKind};
 
 use crate::error::{PlanError, PlanResult};
 use crate::resolver::state::PlanResolverState;
@@ -86,132 +78,5 @@ impl PlanResolver<'_> {
             command,
             return_type,
         })
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    pub(super) fn resolve_python_udf_expr(
-        &self,
-        function: PythonUdf,
-        name: &str,
-        arguments: Vec<Expr>,
-        argument_display_names: &[String],
-        schema: &DFSchemaRef,
-        deterministic: bool,
-        distinct: bool,
-        state: &mut PlanResolverState,
-    ) -> PlanResult<Expr> {
-        use spec::PySparkUdfType;
-
-        let mut scope = state.enter_config_scope();
-        let state = scope.state();
-        state.config_mut().arrow_allow_large_var_types = true;
-
-        let input_types: Vec<DataType> = arguments
-            .iter()
-            .map(|arg| arg.get_type(schema))
-            .collect::<Result<Vec<DataType>, DataFusionError>>()?;
-        let payload = PySparkUdfPayload::build(
-            &function.python_version,
-            &function.command,
-            function.eval_type,
-            &((0..arguments.len()).collect::<Vec<_>>()),
-            &self.config.pyspark_udf_config,
-        )?;
-
-        match function.eval_type {
-            PySparkUdfType::None
-            | PySparkUdfType::GroupedMapPandas
-            | PySparkUdfType::WindowAggPandas
-            | PySparkUdfType::MapPandasIter
-            | PySparkUdfType::CogroupedMapPandas
-            | PySparkUdfType::MapArrowIter
-            | PySparkUdfType::GroupedMapPandasWithState
-            | PySparkUdfType::Table
-            | PySparkUdfType::ArrowTable => Err(PlanError::invalid(format!(
-                "unsupported Python UDF type for common inline UDF: {:?}",
-                function.eval_type
-            ))),
-            PySparkUdfType::Batched => {
-                let udf = PySparkUDF::new(
-                    PySparkUdfKind::Batch,
-                    get_udf_name(name, &payload),
-                    payload,
-                    deterministic,
-                    input_types,
-                    function.output_type,
-                    self.config.pyspark_udf_config.clone(),
-                );
-                Ok(Expr::ScalarFunction(expr::ScalarFunction {
-                    func: Arc::new(ScalarUDF::from(udf)),
-                    args: arguments,
-                }))
-            }
-            PySparkUdfType::ArrowBatched => {
-                let udf = PySparkUDF::new(
-                    PySparkUdfKind::ArrowBatch,
-                    get_udf_name(name, &payload),
-                    payload,
-                    deterministic,
-                    input_types,
-                    function.output_type,
-                    self.config.pyspark_udf_config.clone(),
-                );
-                Ok(Expr::ScalarFunction(expr::ScalarFunction {
-                    func: Arc::new(ScalarUDF::from(udf)),
-                    args: arguments,
-                }))
-            }
-            PySparkUdfType::ScalarPandas => {
-                let udf = PySparkUDF::new(
-                    PySparkUdfKind::ScalarPandas,
-                    get_udf_name(name, &payload),
-                    payload,
-                    deterministic,
-                    input_types,
-                    function.output_type,
-                    self.config.pyspark_udf_config.clone(),
-                );
-                Ok(Expr::ScalarFunction(expr::ScalarFunction {
-                    func: Arc::new(ScalarUDF::from(udf)),
-                    args: arguments,
-                }))
-            }
-            PySparkUdfType::ScalarPandasIter => {
-                let udf = PySparkUDF::new(
-                    PySparkUdfKind::ScalarPandasIter,
-                    get_udf_name(name, &payload),
-                    payload,
-                    deterministic,
-                    input_types,
-                    function.output_type,
-                    self.config.pyspark_udf_config.clone(),
-                );
-                Ok(Expr::ScalarFunction(expr::ScalarFunction {
-                    func: Arc::new(ScalarUDF::from(udf)),
-                    args: arguments,
-                }))
-            }
-            PySparkUdfType::GroupedAggPandas => {
-                let udaf = PySparkGroupAggregateUDF::new(
-                    get_udf_name(name, &payload),
-                    payload,
-                    deterministic,
-                    argument_display_names.to_vec(),
-                    input_types,
-                    function.output_type,
-                    self.config.pyspark_udf_config.clone(),
-                );
-                Ok(Expr::AggregateFunction(expr::AggregateFunction {
-                    func: Arc::new(AggregateUDF::from(udaf)),
-                    params: AggregateFunctionParams {
-                        args: arguments,
-                        distinct,
-                        filter: None,
-                        order_by: vec![],
-                        null_treatment: None,
-                    },
-                }))
-            }
-        }
     }
 }
